@@ -3,32 +3,33 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ant-design/icons": "^4.2.1",
-"@cube-dev/example-wrapper": "^1.1.9",
-    "@cubejs-client/core": "^0.19.37",
-    "@cubejs-client/react": "^0.19.37",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "antd": "^4.3.1",
-    "axios": "^0.19.2",
-    "deck.gl": "^8.1.9",
-    "less": "^3.11.3",
-    "less-loader": "^6.1.1",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-color": "^2.18.1",
+    "@ant-design/icons": "^4.7.0",
+    "@cube-dev/example-wrapper": "^1.1.9",
+    "@cubejs-client/core": "^0.30.29",
+    "@cubejs-client/react": "^0.30.40",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.3.0",
+    "antd": "^4.22.1",
+    "axios": "^0.27.2",
+    "deck.gl": "^8.8.5",
+    "less": "^4.1.3",
+    "less-loader": "^11.0.0",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-color": "^2.19.3",
     "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^16.13.1",
-    "react-map-gl": "^5.2.5",
+    "react-dom": "^18.2.0",
+    "mapbox-gl": "^2.9.2",
+    "react-map-gl": "^7.0.18",
     "react-map-gl-heatmap-overlay": "^1.1.4",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.4",
-    "recharts": "^1.8.5"
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "recharts": "^2.1.13"
   },
   "devDependencies": {
-    "cross-env": "^7.0.2"
+    "cross-env": "^7.0.3"
   },
   "scripts": {
     "start": "cross-env NODE_ENV=development SKIP_PREFLIGHT_CHECK=true react-scripts start",
