@@ -2,7 +2,7 @@ const path = require('path');
 const { renameCategory } = require('./src/rename-category.js');
 
 exports.createPages = ({ actions, graphql }) => {
-  const { createPage } = actions;
+  const { createPage, createRedirect } = actions;
 
   const DocTemplate = path.resolve('src/templates/DocTemplate.tsx');
 
@@ -20,6 +20,7 @@ exports.createPages = ({ actions, graphql }) => {
             scope
             category
             menuOrder
+            redirect_from
           }
         }
       }
@@ -42,6 +43,12 @@ exports.createPages = ({ actions, graphql }) => {
           slug: node.frontmatter.permalink,
         },
       });
+
+      if (Array.isArray(node.frontmatter.redirect_from)) {
+        node.frontmatter.redirect_from.forEach(
+          (from) => createRedirect({ fromPath: from, toPath: node.frontmatter.permalink, isPermanent: true })
+        );
+      }
     });
   });
 };
