@@ -60,7 +60,7 @@ const sample = [
   ]
 ]
 
-export default () => {
+const Heatmap = () => {
   const [viewport, setViewport] = useState({
     latitude: 34,
     longitude: 5,
@@ -117,7 +117,7 @@ export default () => {
         width='100%'
         height='100%'
         mapStyle='mapbox://styles/kalipsik/ckb2fyfqu123n1ilb5yi7uyns/'
-        mapboxApiAccessToken={process.env.REACT_APP_MAPBOX_KEY}
+        mapboxAccessToken={process.env.REACT_APP_MAPBOX_KEY}
       >
         <div className='mapbox__navi'>
           <NavigationControl />
@@ -166,4 +166,4 @@ export default () => {
     </div>);
 }
 
-
+export default Heatmap;
